Separate `main_handler` into separate module.
`main.py` is now only for `Flask` routes.

The real purpose for this was to run tests on the main handler without having to use a Flask test client.
